Preserve significant figures of the original GPS coordinates
The coordinates are stored as EXIF rational numbers which hold
the number of significant figures of the data implicitly.  Using
a fixed rational denominator of 10000 for altitude (as was
previously done, for example) implies a precision of 0.1
millimeters, which is more what GPS technology can provide (and
is well beyond what consumer units can do).  The denominator is
now scaled to the number of decimal places in the original XML
file.  Since at least some GPS units will specify the number of
decimal points in relation to the confidence of the location
reading, this preserves the confidence level in the EXIF tags.