Erase any GPS tags before adding new ones
This prevents duplicate GPS tags in the final file, which could
happen if the same tags already existed in the original file.
Correlation would normally abort with a "GPS Already Present"
error were a file already geotagged, but since only the presence
of the GPSLatitude tag is checked, others might still exist
undetected.